+++ spec/integration/taskrouter/v1/workspace/worker/worker_channel_spec.rb
@@ -35,9 +35,9 @@ describe 'WorkerChannel' do
               "key": "channels",
               "next_page_url": null,
               "page": 0,
-              "page_size": 1,
+              "page_size": 50,
               "previous_page_url": null,
-              "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels"
+              "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels?PageSize=50&Page=0"
           },
           "channels": [
               {
@@ -46,12 +46,12 @@ describe 'WorkerChannel' do
                   "available": true,
                   "available_capacity_percentage": 100,
                   "configured_capacity": 1,
-                  "date_created": "2014-05-14T10:50:02Z",
-                  "date_updated": "2014-05-15T16:03:42Z",
+                  "date_created": "2016-04-14T17:35:54Z",
+                  "date_updated": "2016-04-14T17:35:54Z",
                   "sid": "WCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                   "task_channel_sid": "TCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                   "task_channel_unique_name": "default",
-                  "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/WRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                  "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/WCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                   "worker_sid": "WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                   "workspace_sid": "WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
               }
@@ -77,9 +77,9 @@ describe 'WorkerChannel' do
               "key": "channels",
               "next_page_url": null,
               "page": 0,
-              "page_size": 1,
+              "page_size": 50,
               "previous_page_url": null,
-              "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels"
+              "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels?PageSize=50&Page=0"
           },
           "channels": []
       }
@@ -109,7 +109,7 @@ describe 'WorkerChannel' do
     ))).to eq(true)
   end
 
-  it "receives fetch responses" do
+  it "receives fetch_sid responses" do
     @holodeck.mock(Twilio::Response.new(
         200,
       %q[
@@ -119,12 +119,12 @@ describe 'WorkerChannel' do
           "available": true,
           "available_capacity_percentage": 100,
           "configured_capacity": 1,
-          "date_created": "2014-05-14T10:50:02Z",
-          "date_updated": "2014-05-15T16:03:42Z",
+          "date_created": "2016-04-14T17:35:54Z",
+          "date_updated": "2016-04-14T17:35:54Z",
           "sid": "WCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "task_channel_sid": "TCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "task_channel_unique_name": "default",
-          "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/WRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+          "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/WCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "worker_sid": "WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "workspace_sid": "WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
       }
@@ -164,8 +164,8 @@ describe 'WorkerChannel' do
           "available": true,
           "available_capacity_percentage": 100,
           "configured_capacity": 3,
-          "date_created": "2014-05-14T10:50:02Z",
-          "date_updated": "2014-05-15T16:03:42Z",
+          "date_created": "2016-04-14T17:35:54Z",
+          "date_updated": "2016-04-14T17:35:54Z",
           "sid": "WCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "task_channel_sid": "TCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "task_channel_unique_name": "default",