+++ spec/integration/verify/v2/service/webhook_spec.rb
@@ -47,6 +47,7 @@ describe 'Webhook' do
           "webhook_method": "POST",
           "webhook_url": "https://owlbank.twilio.com",
           "status": "enabled",
+          "version": "v2",
           "date_created": "2015-07-30T20:00:00Z",
           "date_updated": "2015-07-30T20:00:00Z"
       }
@@ -91,6 +92,7 @@ describe 'Webhook' do
           "webhook_method": "POST",
           "webhook_url": "https://owlbank.twilio.com",
           "status": "disabled",
+          "version": "v2",
           "date_created": "2015-07-30T20:00:00Z",
           "date_updated": "2015-07-30T20:00:00Z"
       }
@@ -162,6 +164,7 @@ describe 'Webhook' do
           "webhook_method": "POST",
           "webhook_url": "https://owlbank.twilio.com",
           "status": "enabled",
+          "version": "v2",
           "date_created": "2015-07-30T20:00:00Z",
           "date_updated": "2015-07-30T20:00:00Z"
       }
@@ -233,6 +236,7 @@ describe 'Webhook' do
                   "webhook_method": "POST",
                   "webhook_url": "https://owlbank.twilio.com",
                   "status": "enabled",
+                  "version": "v2",
                   "date_created": "2015-07-30T20:00:00Z",
                   "date_updated": "2015-07-30T20:00:00Z"
               }